refactor: refactor in to src layout
