+++ pyproject.toml
@@ -47,7 +47,6 @@ ignore = [
     "TRY003",  # Exception with string (vs. defining custom exception)
     "PLR2004", # Magic number
 ]
-include = ["*stellarmesh.py"]
 target-version = "py39"
 
 [tool.ruff.pydocstyle]
@@ -57,5 +56,4 @@ convention = "google"
 pythonVersion = "3.9"
 
 [tool.black]
-include = '.*stellarmesh\.py$'
 target-version = ["py39", "py310", "py311"]
