+++ src/stellarmesh/__init__.py
@@ -1,0 +1,4 @@
+"""GMSH wrapper and DAGMC geometry creator."""
+from geometry import Geometry  # noqa: F401
+from mesh import Mesh  # noqa: F401
+from moab import MOABModel  # noqa: F401
