+++ src/stellarmesh/_core.py
@@ -1,0 +1,3 @@
+import logging
+
+logger = logging.getLogger(__name__)
