+++ src/stellarmesh/mesh.py
@@ -1,0 +1,158 @@
+"""Stellarmesh mesh.
+
+name: mesh.py
+author: Alex Koen
+
+desc: Mesh class wraps Gmsh functionality for geometry meshing.
+"""
+import logging
+import tempfile
+from typing import Optional
+
+import gmsh
+
+from ._core import logger
+from .geometry import Geometry
+
+
+class Mesh:
+    """A Gmsh mesh.
+
+    As gmsh allows for only a single process, this class provides a context manager to
+    set the gmsh api to operate on this mesh.
+    """
+
+    _mesh_filename: str
+
+    def __init__(self, mesh_filename: Optional[str] = None):
+        """Initialize a mesh from a .msh file.
+
+        Args:
+            mesh_filename: Optional .msh filename. If not provided defaults to a
+            temporary file. Defaults to None.
+        """
+        if not mesh_filename:
+            with tempfile.NamedTemporaryFile(suffix=".msh", delete=False) as mesh_file:
+                mesh_filename = mesh_file.name
+        self._mesh_filename = mesh_filename
+
+    def __enter__(self):
+        """Enter mesh context, setting gmsh commands to operate on this mesh."""
+        if not gmsh.is_initialized():
+            gmsh.initialize()
+
+        gmsh.option.setNumber(
+            "General.Terminal",
+            1 if logger.getEffectiveLevel() <= logging.INFO else 0,
+        )
+        gmsh.open(self._mesh_filename)
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        """Cleanup (finalize) gmsh."""
+        gmsh.finalize()
+
+    def _save_changes(self, *, save_all: bool = True):
+        gmsh.option.set_number("Mesh.SaveAll", 1 if save_all else 0)
+        gmsh.write(self._mesh_filename)
+
+    def write(self, filename: str, *, save_all: bool = True):
+        """Write mesh to a .msh file.
+
+        Args:
+            filename: Path to write file.
+            save_all: Whether to save all entities (or just physical groups). See
+            Gmsh documentation for Mesh.SaveAll. Defaults to True.
+        """
+        with self:
+            gmsh.option.set_number("Mesh.SaveAll", 1 if save_all else 0)
+            gmsh.write(filename)
+
+    @classmethod
+    def mesh_geometry(
+        cls,
+        geometry: Geometry,
+        min_mesh_size: float = 50,
+        max_mesh_size: float = 50,
+    ):
+        """Mesh solids with Gmsh.
+
+        See Gmsh documentation on mesh sizes:
+        https://gmsh.info/doc/texinfo/gmsh.html#Specifying-mesh-element-sizes
+
+        Args:
+            geometry: Geometry to be meshed.
+            mesh_filename: Optional filename to store .msh file. Defaults to None.
+            min_mesh_size: Min mesh element size. Defaults to 50.
+            max_mesh_size: Max mesh element size. Defaults to 50.
+        """
+        logger.info(f"Meshing solids with mesh size {min_mesh_size}, {max_mesh_size}")
+
+        with cls() as mesh:
+            gmsh.model.add("stellarmesh_model")
+
+            material_solid_map = {}
+            for s, m in zip(geometry.solids, geometry.material_names):
+                dim_tags = gmsh.model.occ.import_shapes_native_pointer(
+                    s.wrapped._address()
+                )
+                if dim_tags[0][0] != 3:
+                    raise TypeError("Importing non-solid geometry.")
+
+                solid_tag = dim_tags[0][1]
+                if m not in material_solid_map:
+                    material_solid_map[m] = [solid_tag]
+                else:
+                    material_solid_map[m].append(solid_tag)
+
+            gmsh.model.occ.synchronize()
+
+            for material, solid_tags in material_solid_map.items():
+                gmsh.model.add_physical_group(3, solid_tags, name=f"mat:{material}")
+
+            gmsh.option.set_number("Mesh.MeshSizeMin", min_mesh_size)
+            gmsh.option.set_number("Mesh.MeshSizeMax", max_mesh_size)
+            gmsh.model.mesh.generate(2)
+
+            mesh._save_changes(save_all=True)
+            return mesh
+
+    def render(
+        self,
+        output_filename: Optional[str] = None,
+        rotation_xyz: tuple[float, float, float] = (0, 0, 0),
+        normals: int = 0,
+        *,
+        clipping: bool = True,
+    ) -> str:
+        """Render mesh as an image.
+
+        Args:
+            output_filename: Optional output filename. Defaults to None.
+            rotation_xyz: Rotation in Euler angles. Defaults to (0, 0, 0).
+            normals: Normal render size. Defaults to 0.
+            clipping: Whether to enable mesh clipping. Defaults to True.
+
+        Returns:
+            Path to image file, either passed output_filename or a temporary file.
+        """
+        with self:
+            gmsh.option.set_number("Mesh.SurfaceFaces", 1)
+            gmsh.option.set_number("Mesh.Clip", 1 if clipping else 0)
+            gmsh.option.set_number("Mesh.Normals", normals)
+            gmsh.option.set_number("General.Trackball", 0)
+            gmsh.option.set_number("General.RotationX", rotation_xyz[0])
+            gmsh.option.set_number("General.RotationY", rotation_xyz[1])
+            gmsh.option.set_number("General.RotationZ", rotation_xyz[2])
+            if not output_filename:
+                with tempfile.NamedTemporaryFile(
+                    delete=False, mode="w", suffix=".png"
+                ) as temp_file:
+                    output_filename = temp_file.name
+
+            try:
+                gmsh.fltk.initialize()
+                gmsh.write(output_filename)
+            finally:
+                gmsh.fltk.finalize()
+            return output_filename
